Refactoring <App />
